InvPro/requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.11.0
future==0.18.2
mysqlclient @ file:///D:/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==7.1.2
pytz==2020.1
serializers==0.2.4
sqlparse==0.3.1
views==0.3
